Send data in request body for POST, PUT
Previously, when issuing a POST (or PUT) request, the payload was sent
within the query string. Even though there to my knowledge are no
theoretical limits on query string length in the HTTP specification, in
practice web servers have their respective quirks, often limiting the
length around 4k characters [0].

The Yammer servers seem to have a similar limit, making e.g.
announcements that break ~4k characters yield "HTTP 404" when issuing
the request (why it is "404 Not found" and not e.g. "413 Entity too
large" is an open question).

POST (and PUT) requests that often carry large amounts of data for this
reason generally carry their payload within the request body. This is
what this commit implements.

The tests previously explicitly checked for finding data in the query
string, but have been modified to look for the sent data in either the
query string or the request body, which should be a reasonable
acceptance test for these cases.

[0]: <https://stackoverflow.com/a/812962>